feat: finish 4.21 task

+++ controllers/blogs.js
@@ -1,8 +1,8 @@
 /* eslint-disable import/no-extraneous-dependencies */
 const blogsRouter = require('express').Router();
-const jwt = require('jsonwebtoken');
 const Blog = require('../models/post');
 const User = require('../models/users');
+const jwt = require('jsonwebtoken');
 
 // get blogs
 blogsRouter.get('/', async (request, response) => {
@@ -77,10 +77,29 @@ blogsRouter.put('/:id', async (request, response) => {
   }
 });
 
-// delete - task 4.13
+// delete - task 4.13 - token expansion task 4.21
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const blog = await Blog.findById(request.params.id);
+
+  // authorization and verification
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  if (!decodedToken.id) {
+    return response
+      .status(401)
+      .json({ error: 'invalid token, cant delete post' });
+  }
+
+  const user = await User.findById(decodedToken.id);
+  // console.log(`${blog.user.toString()} must equeal ${user.id.toString()} `);
+
+  const isValidUser = blog.user.toString() === user.id.toString();
+
+  if (isValidUser) {
+    await Blog.findByIdAndRemove(request.params.id);
+  }
+
+  return response.status(204).end();
 });
 
 module.exports = blogsRouter;
